Run Prometheus exporter as a separate service
